requirements.txt
django==4.2.11
djangorestframework==3.15.1

# CORS Headers
django-cors-headers==3.10.1